+++ packages/contracts/package.json
@@ -1,5 +1,6 @@
 {
   "name": "@machina-finance/contracts",
+  "type": "module",
   "version": "0.0.1",
   "description": "",
   "main": "index.js",
@@ -12,6 +13,7 @@
   "dependencies": {
     "@fleet-sdk/mock-chain": "^0.2.2",
     "@fleet-sdk/core": "^0.2.2",
+    "@fleet-sdk/crypto": "^0.2.2",
     "@fleet-sdk/compiler": "^0.2.2",
     "@fleet-sdk/common": "^0.2.2",
     "@fleet-sdk/serializer": "^0.2.2"
