+++ packages/contracts/tsconfig.json
@@ -1,0 +1,6 @@
+{
+  "extends": "@tsconfig/recommended/tsconfig.json",
+  "compilerOptions": {
+    "types": ["bun-types"]
+  }
+}
